requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pydantic
extract-thinker
litellm
openpyxl
python-dotenv
docx